chore: move PieChartConfigTabs to folder
